refactor: mysql dump by GetDatabaseDefinition
